nfcrypto: a real function, proper startup.
Convert the nfcrypto.js code, which was supposed to be executed at the
browser's top level, to a function that may be invoked at any time.
Thus, openpgp.init() may actually attempt to initialize the NfWebCrypto
plugin now.

To this end, now that the code may be executed much, much later than
window.onLoad(), just attempt to run the onLoad() function as part of
the openpgp_nfcrypto_init() function so that the plugin is really set
up properly.