Generate (and throw away) an encryption keypair.
This is just a demonstration of how the encryption subkey's
generation will be performed and that does succeed using
either the "owncrypto" or "nfwebcrypto" WebCrypto provider.

Actual inclusion of the encryption subkey into the private and
public key material exported by openpgp.generate_key_pair() is
currently subject to further discussion.