Stash WebCrypto keys into the keyring: almost done.
When writing out a private key, check if it's extractable and,
if yes, write it in a special format using the reserved value 100
for the string-to-key symmetric encryption algorithm.  For the present,
stash the WebCrypto provider name, the key name and the key ID as
MPIs representing not numbers, but strings.